Fix migration when there is no printerlogs/networkportmetrics

+++ install/migrations/update_9.5.x_to_10.0.0/native_inventory.php
@@ -534,12 +534,14 @@ if (!$DB->tableExists('glpi_printerlogs')) {
             )
         );
         $to_preserve_result = $DB->query($to_preserve_sql->getValue())->fetch_all(MYSQLI_ASSOC);
-        $DB->delete(
-            'glpi_printerlogs',
-            [
-                'NOT' => ['id' => array_column($to_preserve_result, 'id')]
-            ]
-        );
+        if (!empty($to_preserve_result)) { // If there is no entries to preserve, it means that table is empty, and nothing has to be deleted
+            $DB->delete(
+                'glpi_printerlogs',
+                [
+                    'NOT' => ['id' => array_column($to_preserve_result, 'id')]
+                ]
+            );
+        }
         $migration->addKey('glpi_printerlogs', ['printers_id', 'date'], 'unicity', 'UNIQUE');
     }
 }
@@ -612,12 +614,14 @@ if (!$DB->tableExists('glpi_networkportmetrics')) {
             )
         );
         $to_preserve_result = $DB->query($to_preserve_sql->getValue())->fetch_all(MYSQLI_ASSOC);
-        $DB->delete(
-            'glpi_networkportmetrics',
-            [
-                'NOT' => ['id' => array_column($to_preserve_result, 'id')]
-            ]
-        );
+        if (!empty($to_preserve_result)) { // If there is no entries to preserve, it means that table is empty, and nothing has to be deleted
+            $DB->delete(
+                'glpi_networkportmetrics',
+                [
+                    'NOT' => ['id' => array_column($to_preserve_result, 'id')]
+                ]
+            );
+        }
         $migration->addKey('glpi_networkportmetrics', ['networkports_id', 'date'], 'unicity', 'UNIQUE');
     }
 }
